% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_perrrfect.R
\name{build_filter}
\alias{build_filter}
\title{This function can be used to build predefined filters.}
\usage{
build_filter(kernel_type, kernel_size)
}
\arguments{
\item{kernel_type}{A string which accepts following values
blur: Used to blur the picture
sharpen: Used to increase the sharpness of the image
More options will be added as enhancements}

\item{kernel_size}{An integer determining the filter size.}
}
\value{
A matrix of size kernel_size * kernel_size representing the filter.
}
\description{
This function can be used to build predefined filters.
}
\examples{
\donttest{
build_filter("blur", 3)
build_filter("sharpen", 7)
}
}
