% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression_perrrfect.R
\name{check_values}
\alias{check_values}
\title{This function is to be used in conjunction with compression_pyfect.}
\usage{
check_values(img, kernel_size)
}
\arguments{
\item{img}{A cimg type object of 4 dimensions representing a greyscale or colour image.}

\item{kernel_size}{The size of the length and width of the kernel to be passed over the image. Default: 2}
}
\value{

}
\description{
This function checks that the img and kernel size are valid inputs
and stops with an error if not.
}
\examples{
\donttest{
img <- imager::as.cimg(imager::boats)
check_values(img, kernel_size=2)
}
}
