% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression_perrrfect.R
\name{compression_perrrfect}
\alias{compression_perrrfect}
\title{This function uses a lossy pooling algorithm to compress an image.}
\usage{
compression_perrrfect(img, kernel_size = 2, pooling_function = "max")
}
\arguments{
\item{img}{A cimg type object of 4 dimensions representing a greyscale or colour image.}

\item{kernel_size}{The size of the length and width of the kernel to be passed over the image. Default: 2}

\item{pooling_function}{The pooling algorithm to be used. Three options: "max", "min", and "mean". Default: "max"}
}
\value{
A cimg type image of 4 dimensions representing a greyscale or colour image.
}
\description{
The function can be applied to a greyscale or colour image. The user passes
a cimg type object (from the imager package) which is to be compressed and the resulting
compressed cimg object is returned. The user can also specify the pooling algorithm to be
used and the size of the kernel to apply over the image.
}
\examples{
img <- imager::as.cimg(imager::boats)
compression_perrrfect(img, kernel_size=2, pooling_function="max")
}
