% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_perrrfect.R
\name{filter_perrrfect}
\alias{filter_perrrfect}
\title{This function can be used to apply predefined or custom filters on an image.}
\usage{
filter_perrrfect(
  image,
  filter_type = "blur",
  filter_size = 3,
  custom_filter = NULL
)
}
\arguments{
\item{image}{A cimg object to representing the image}

\item{filter_type}{One of the following values:
"blur": Used to blur the picture
"sharpen": Used to increase the sharpness of the image
"custom": Allows users to use their own filter
More options will be added as enhancements}

\item{filter_size}{An integer determining the filter size.
This is used if the filter_type is not custom. Default: 3}

\item{custom_filter}{A 2d matrix that allows users to pass their own filter.
This is only used if the users select filter_type = "custom"}
}
\value{
A cimg object representing the transformed image.
}
\description{
The function can be applied on images represented by cimg object. The users can
choose from predefined filters or can create their new filters. This can be used
for various purposes like entertainment application or visualization of convolutional
neural network.
}
\examples{
filter_perrrfect(imager::boats, filter_type="blur", filter_size=21L, custom_filter=NULL)
filter_perrrfect(imager::boats, filter_type="custom", custom_filter=matrix(0.01, 42, 21))
}
