% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_property_perrrfect.R
\name{get_property_perrrfect}
\alias{get_property_perrrfect}
\title{This function will get the image property of a image.}
\usage{
get_property_perrrfect(image)
}
\arguments{
\item{image}{imager::cimg image to obtain image information}
}
\value{
a list of image dimensions, pixels, and each channel's mean and median
}
\description{
This function will get the image property of a image.
}
\examples{
\dontrun{
get_property_perrrfect(imager::boats)
}
}
