% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression_perrrfect.R
\name{pool_band}
\alias{pool_band}
\title{This function is to be used in conjunction with compression_pyfect
and compresses a single colour band of an image.}
\usage{
pool_band(img, kernel_size, pooling_func)
}
\arguments{
\item{img}{A 'matrix' 'array' type object of 2 dimensions representing a color band
of either a greyscale image or a colour image.}

\item{kernel_size}{The size of the length and width of the kernel to be passed over the image. Default: 2}

\item{pooling_func}{The pooling algorithm to be used. Three options: "max", "min", and "mean". Default: "max"}
}
\value{
A cimg type image of 4 dimensions representing a greyscale or colour image.
}
\description{
The function applies a lossy pooling algorithm to compress the specified
colour band and the resulting cimg object is returned.
}
\examples{
\donttest{
img <- imager::as.cimg(imager::boats)
pool_band(img[,,,1], kernel_size=2, pooling_func="max")
}
}
