% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_perrrfect.R
\name{rotation_perrrfect}
\alias{rotation_perrrfect}
\title{rotation_perrrfect rotates an image in matrix form by
either 90, 180, 270, or 360 degrees.}
\usage{
rotation_perrrfect(image, n_rot)
}
\arguments{
\item{image}{imager::cimg of pixel values to be rotated}

\item{n_rot}{Integer number of 90-degree rotations to return}
}
\value{
imager::cimg of rotated pixel values
}
\description{
rotation_perrrfect rotates an image in matrix form by
either 90, 180, 270, or 360 degrees.
}
\examples{
\dontrun{
rotation_perrrfect(imager::boats, 1)
}
}
